[IE Myriad] Move MyriadDetectionOutput tests to the new test infrastructure
